Move in game console to UI library
