+++ src/openrct2-ui/input/Input.cpp
@@ -25,8 +25,12 @@
 #include <openrct2/interface/Console.h>
 #include <openrct2/paint/VirtualFloor.h>
 #include <openrct2-ui/windows/Window.h>
+#include "../interface/InGameConsole.h"
+#include "../UiContext.h"
 #include "KeyboardShortcuts.h"
 #include "Input.h"
+
+using namespace OpenRCT2::Ui;
 
 static void input_handle_console(sint32 key)
 {
@@ -54,7 +58,8 @@ static void input_handle_console(sint32 key)
     }
     if (input != CONSOLE_INPUT_NONE)
     {
-        console_input(input);
+        auto& console = GetInGameConsole();
+        console.Input(input);
     }
 }
 
@@ -129,7 +134,8 @@ void input_handle_keyboard(bool isTitle)
         return;
     }
 
-    if (!console_is_open())
+    auto& console = GetInGameConsole();
+    if (!console.IsOpen())
     {
         if (!isTitle)
         {
@@ -188,14 +194,14 @@ void input_handle_keyboard(bool isTitle)
         // Reserve backtick for console
         if (key == SDL_SCANCODE_GRAVE)
         {
-            if ((gConfigGeneral.debugging_tools && !context_is_input_active()) || console_is_open())
+            if ((gConfigGeneral.debugging_tools && !context_is_input_active()) || console.IsOpen())
             {
                 window_cancel_textbox();
-                console_toggle();
+                console.Toggle();
             }
             continue;
         }
-        else if (console_is_open())
+        else if (console.IsOpen())
         {
             input_handle_console(key);
             continue;
