+++ src/openrct2/Context.cpp
@@ -175,6 +175,11 @@ namespace OpenRCT2
             return gExitCode;
         }
 
+        void WriteLine(const std::string &s) override
+        {
+            _stdInOutConsole.WriteLine(s);
+        }
+
         /**
          * Causes the OpenRCT2 game loop to finish.
          */
@@ -793,8 +798,8 @@ namespace OpenRCT2
 
             twitch_update();
             chat_update();
-            console_update();
             _stdInOutConsole.ProcessEvalQueue();
+            _uiContext->Update();
         }
 
         /**
