+++ src/openrct2/paint/Painter.cpp
@@ -52,7 +52,7 @@ void Painter::Paint(IDrawingEngine * de)
 
         update_palette_effects();
         chat_draw(dpi);
-        console_draw(dpi);
+        _uiContext->Draw(dpi);
 
         if ((gScreenFlags & SCREEN_FLAGS_TITLE_DEMO) && !title_should_hide_version_info())
         {
